refactor: domain 패키지 재정의 * domain.station 패키지 삭제
