+++ src/main/java/subway/api/controller/StationController.java
@@ -4,7 +4,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.api.dto.StationRequest;
 import subway.api.dto.StationResponse;
-import subway.domain.station.service.StationService;
+import subway.domain.service.StationService;
 
 import java.net.URI;
 import java.util.List;
