+++ src/main/java/subway/domain/repository/StationRepository.java
@@ -1,7 +1,7 @@
-package subway.domain.station.repository;
+package subway.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.domain.station.entity.Station;
+import subway.domain.entity.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
 }
