+++ src/main/java/subway/domain/service/StationService.java
@@ -1,11 +1,11 @@
-package subway.domain.station.service;
+package subway.domain.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import subway.api.dto.StationRequest;
 import subway.api.dto.StationResponse;
-import subway.domain.station.entity.Station;
-import subway.domain.station.repository.StationRepository;
+import subway.domain.entity.Station;
+import subway.domain.repository.StationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
